requirements.txt
fastapi
uvicorn
pydantic
gitpython
python-dotenv
requests